Solve day 1 part 1.
I initially wanted to parse the input from the file input.txt, but
then I would have had to figure out so many things:

* How do I get the contents of a file in zig?
* Wait, how do I allocate memory for the file name?
* Every function in std.fs seems to want an absolute path, how do I
  get the current path?
* Should I be looking in std.fs, std.io, std.os, elsewhere?
* Do I want a BufferedReader? What's a ReaderType?
* How do you parse an integer, is that in std.fmt?

Instead, I copy-pasted the input as a constant global array and let
the compiler do the work for me.

Similarly, I initially thought I'd use a hash-map to keep track of
the numbers already seen in the list, but a smallish array of bools
is easier to figure out.

Anyway, it works!

$ zig build run
935419

On to part 2.