Solve day 1 part 2.
Things I learned for this part:

* [ziglearn.org] exists, which I sorely needed,
* `for` loops can iterate over (value, index) pairs,
* Zig has array slices, which use the syntax `a[m..n]` for the slice
  including `a[m]` up to `a[n-1]`, and
* Zig has optionals, which use the syntax `?T` for any type `T`.

I already knew that [Rosetta Code] and [the #zig irc backlog] exist,
and I'm sure they will be useful in the future too.

[ziglearn.org]: <https://ziglearn.org/>
[Rosetta Code]: <https://rosettacode.org/wiki/Category:Zig>
[the #zig irc backlog]: <https://freenode.irclog.whitequark.org/zig/>

For this part, instead of recording whether a number has been seen
in an array of bools, I record whether a sum of two numbers has been
seen. Actually, when a sum of two numbers has been seen, I record the
product instead of just true/false. (If a sum of two numbers is seen
more than once, I glibly overwrite the product with a new one.)

Zig has already detected integer overflows and out-of-bounds array
accesses in my code, which is good. It doesn't check statically, so
I can't know the code is ok before running it, but at least I can
know that the code happened to be ok after running it (that is, no
silent errors corrupting my data).

$ zig build run
49880012

Day 1 is a success!