Solve day 2 part 1.
Today I learned:

* There's a nice list of projects in Zig at [Awesome Zig], which may
  include useful examples of code and useful libraries.

* `for` loops are only for slices/arrays, not custom iterators, but
  `while` loops are [quite featureful]! In fact it seems like the
  language creator has been tempted to [delete for loops] from the
  language and only have while loops.

* The language reference says that "Zig programmers must always be
  able to answer the question: [Where are the bytes]?" and this is
  why you have to choose an allocator, but even after reading some
  of the source code I really don't know where the 4KB buffer for
  bufferedReader lives.

* I can read a file after all! Thanks to chapter 2 of ziglearn.org.

[Awesome Zig]: <https://github.com/nrdmn/awesome-zig>
[quite featureful]: <https://ziglang.org/documentation/0.7.0/#while>
[delete for loops]: <https://github.com/ziglang/zig/issues/3110#issuecomment-584424303>
[Where are the bytes]: <https://ziglang.org/documentation/0.7.0/#Memory>

For the solution, I got frustrated with ArrayList and the helper
functions in std.io.Reader and std.fmt, so I decided to just do a
finite state machine which reads one character at a time. I think
the result is not so bad, but I have to admit that I'm surprised it
gave the correct answer on the first run, because it would have been
extremely easy to forget updating the parsing_state or some other
trivial error in the state machine.

$ zig build run
643

On to part 2, and let's hope it doesn't need much more state.