Solve day 4 part 1.
Today I decided to try out a concrete example of one of Zig's more
interesting features: compile-time evaluation. My code is written as
a triply-nested loop, over:

1. passport records in the input file, then
2. key-value pairs in each passport record, then
3. each of the seven required fields, to see if the current
   key-value pair matches it).

But, I wrote the inner loop (layer 3) as an [inline for] loop, so the
compiled code only contains a doubly-nested loop (layers 1 and 2),
with the inner loop unrolled as a sequence of seven `if` statements.
This is possible because the list of required fields is fixed and
known at compile-time, even though this requires running some code
(to add a ":" character at the end of each field prefix).

[inline for]: <https://ziglang.org/documentation/0.7.0/#inline-for>

Neat!

$ zig build run
206

On to part 2.