Solve day 5 part 2.
I thought this would require sorting or searching in a fancier way,
in which case std.sort has plenty of useful functions, but a linear
pass over the input followed by a linear pass over 1000 bools is just
fine. Also I really like std.mem.indexOf (and friends), because it
deals equally well with strings or arrays of anything.

I don't know if it's because of Zig nudging me or the care that went
into the puzzles' construction or what, but I like how straight-
-forward my code ends up being for all the solutions so far.

$ zig build run
640