Solve day 6 part 1.
Today I learned about multiline string literals in Zig. I've never
seen this particular syntax for it in another language, but it's
kind of cute!

Coming from Python, I've noticed that my first thought when seeing
the puzzles so far is something like "I'll use a set!", and I'm
sure Zig has a decent set implementation in the standard library,
but it also has very comfortable array and slice builtins, and
that's been sufficient.

$ zig build run
6947

On to part 2.