Solve day 6 part 2.
For this part, I decided to have fun with bit fields. It feels kind
of dirty and dangerous but also exciting!

Zig allows signed and unsigned integer types of arbitrary widths,
everything from i0 to i65535 and from u0 to u65535, and I'm dealing
with sets of 26 boolean values (for the questions `a` through `z`),
so I'm using the type u26 as a bit field.

One of the complications is that integer literals in the source
code are of type `comptime_int`, which is something like a bignum,
and Zig doesn't always want to cast that to a specific bit width,
like u26 or u5. For example, you can't do a bitwise `not` on a
comptime_int, because arguably that would be an integer overflow
or something. But you can explicitly cast the comptime_int literal
to a u26 before doing the bitwise `not` so it's fine.

The other complication is that Zig will restrict types on a bit
shift operation. For example, you can only shift a u26 value by a
number of positions which fits in a u5, because shifting a u26 by
more than 26 positions, let alone 31 positions, is undefined
behaviour. But you can ask Zig to do a potentially unsafe integer
case, like I did:

    @intCast(u5, char - 'a')

I think these complications are good, though. Sure, Zig is letting
me do some potentially unsafe operations, but it's making me be
explicit about it, in a way that a code reviewer could easily see
(especially with the use of a linter) and scrutinize more carefully.

$ zig build run
3398