Solve day 7 part 1.
It turns out memory management is hard! In the previous version of
the parser I had a memory leak that I thought I fixed, but instead
I turned it into a use-after-free bug.

(Maybe I should learn Rust instead.)

As for the actual processing of the rules, I'm doing it with a
depth-first search of the graph of which colours contain which
colours, while being careful to note that any colours contained in
a cycle of inclusions cannot follow the luggage rules.

I don't think this is a sufficient description of the algorithm to
be of much help to any potential readers, but I'm too tired to write
more, sorry.

$ zig build run
259

On to part 2.