Solve day 8 part 2.
For about a minute, I tried thinking of a clever way to interpret
the boot code non-deterministically, and whether that would require
backtracking, etc., but then I decided to be lazy and just run the
existing interpreter once for each possible 'fix', resetting all the
`seen` fields in between runs.

The new code is totally straightforward, and it still runs instantly,
as far as I can tell.

$ zig build run
1205

On to day 9!