Solve day 9 part 1.
Back to easy parsing! This time I just inlined the parsing code
right into main(), and gave the test case an array of numbers rather
than text to be parsed.

There might be a more clever way of checking for an invalid number,
but even then the book-keeping involved might not be cheaper than
this simple brute-force approach. In any case, given that part 2
doesn't have anything to do with part 1 algorithmically, I'm glad I
didn't waste any time on being more clever than this.

$ zig build run
36845998

On to part 2.