Solve day 9 part 2.
This time, I learned from the test case that I should read more
carefully (I initially returned first-in-range plus last-in-range,
rather than min-in-range plus max-in-range), thankfully very
quickly, with the help of std.debug.print().

For this part I'm pleased that I went with a slightly clever
algorithm, where the contiguous slice of numbers under consideration
either grows or shrinks at each step, and something very close to
the minimal amount of work is done to find the target slice.

On to day 10!