Solve day 10 part 2.
Technically, the recurrence for `ways` only needs to remember the
previous 3 elements of the array, rather than all the previous
elements of the array, but I decided to splurge and spend 1 KB of
memory rather than deal with the necessary book-keeping.

Based on the warning in the puzzle text that the answer would be
"more than a trillion", I opted for using 64-bit integers. And
indeed, the answer would not have fit in 32 bits:

$ zig build run
84627647627264

On to day 11.