Solve day 13 part 1.
Yay, modular arithmetic! Also, the puzzle input is small enough that
I decided to regress even further in terms of parsing and just
include it in the source file as a constant and an array.

I initially wrote the modulo as `(-earliest) % id`, but I'm very
happy with the error message I got as a result:

> signed integers and floats must use @rem or @mod

These two functions have different behaviour with respect to sign.
In both cases, the second argument must be positive and nonzero,
and the results fall in the following ranges:

    -b < @rem(a, b) <= 0 <= @mod(a, b) < b

$ zig build run
4808

On to part 2.