Solve day 13 part 2.
Hurray for the Chinese Remainder Theorem and Bézout's identity and
the Extended Euclidean Algorithm! My implementation uses variable
names with more than one letter each, and spells out the notable
invariants that it maintains.

I'm glad I made my implementation generic with respect to the width
of the integers used, because I overflowed 64 bits, and was able to
easily swap in 128 bits. (I also got a slight thrill at the thought
that I could have just as easily swapped in 65535 bits.)

$ zig build run
741745043105674

On to day 14.