Solve day 14 part 2.
This was short, worked on the first try, and involved a nice little
bit hack:

```zig
var float = float_mask;
while (true) {
    try memory.put(address | float, value);
    if (float == 0) break;
    float = (float - 1) & float_mask;
}
```

This loop iterates (backwords) over all numbers whose 1 bits are
contained in the 1 bits of `float_mask`. This works because the
subtraction will borrow across stretches of 0 bits, and the and'ing
will restore the correct stretches of 0 bits.

(For a minute when I saw the description of part 2, I was worried/
hopeful that maybe some masks would have so many floating bits that
iterating over all possibilities would be too slow, which might have
been solved with a clever data structure, but the worst mask in my
input only has 9 floating bits, which is pretty small.)

$ zig build run
3905642473893

On to day 15.