Solve day 15 part 2.
Somewhat predictably, just changing 2020 to 30_000_000 in the code
for part 1 gave a program which runs for way too long. The obvious
culprit is the backwards search through the spoken history for a
repeat of the last spoken number. In the worst case, this runs
through the entire history, but even in the best case, if the
number 3544142 is spoken (for example), that means there was a
search of length 3544142. If this example is typical, that means
there's a million-fold speedup on the table if the search can be
skipped.

So, for this part, I store for each possible number (they're all
going to be < 30_000_000) the last time it was spoken, if ever.
This runs much faster, almost instantly.

Fun observation: I tried having the giant array being a global
constant, so that I wouldn't even have to do any memory allocation,
but that caused the Zig compiler to have a weird kind of seizure for
several seconds before succeeding at compiling. So instead I just do
one memory allocation at run time for the giant array, without
bothering with the usual ArenaAllocator that can handle resizing etc.

$ zig build run
6428

On to day 16.