Solve day 17 part 1.
Zig is definitely not an array programming language. It can do multi-
dimensional arrays with compile-time known dimensions, but if the
dimensions are only run-time known, you have to use slices, and
multi-dimensional slices are a huge pain. If part 2 requires
dynamically-sized multi-dimensional arrays, I swear I'll resort to
doing index arithmetic to simulate multiple dimensions with a single-
dimensional array.

In the end, this solution is very much hardcoded to the particulars
of my own input. I'm not particularly happy with the result.

$ zig build run
271

On to part 2.