Solve day 18 part 1.
Another day, another stateful parser. Because [recursion in Zig] is
a complicated topic, I decided to make an explicit stack to track
the partially evaluated expressions that come up when parentheses
show up.

[recursion in Zig]: <https://ziglang.org/documentation/0.7.0/#Recursion>

And instead of allocating a new stack every time, I decided to reuse
the same stack for every expression evaluation (while asserting that
it starts empty). There's some very nice collaboration between
ArrayLists and Allocators in the Zig standard library:

* When an ArrayList runs out of space, it asks the Allocator for
  more memory, of course, but it only specifies **at least** how
  much.

* The Allocator may increase the allocation in-place (I'm not sure
  if the current implementation of PageAllocator does this), in
  which case there's no need to copy the old contents.

* In any case, the Allocator replies with the amount of memory it
  actually allocated, which it often is in the case of PageAllocator,
  because it gives out allocations in multiples of std.mem.page_size,
  which is 4 KB on Linux.

* With this information, ArrayList accepts the extra capacity and
  can automatically make use of it later.

I'm also quite happy that I initially wrote all of my code with the
number type as `const Num = u32;` with `Num` in all the function
signatures and so on, because that turned out to be too small.
Thankfully Zig detected the overflow at runtime, and I was able to
change it to change only 2 characters to fix the problem.

$ zig build run
209335026987

On to part 2.