Solve day 18 part 2.
I'm pretty happy with how this ended up! This combined parsing/
evaluating method wouldn't scale to a more complicated grammar, but
the grammar we actually have here is very simple.

Also, because multiplication and addition have identity elements on
the left (that is, 1*x=x and 0+x=x for all x), we can do a neat
trick: the parsing state (ignoring parentheses for now) is always
equivalent to the state after parsing something of the form
"x * y +", and expecting a number-like thing, or
"x * y", and expecting an operator-like thing. Even at the start of
a line or right after an opening parenthesis, because this is
equivalent to the state after "1 * 0 +".

$ zig build run
33331817392479

I might take a break now and continue after Christmas.