Solve day 20 part 1.
Ok, this was kind of fun! Parsing the input was not too hard, and
then I didn't actually assemble the picture, I just looked for the
corners by looking at tiles which have two unique edges.

I did note that there are 144 tiles in my input though, so I at least
know that the final picture should be 12x12.

$ zig build run
13983397496713

On to part 2, which will almost certainly require assembly :p