Solve day 20 part 2.
The code is a mess but part 2 was fun too! I decided to have fun
with some of the names, too: the tiles flip and flop until they're
in the right orientation, and the monster does a little dance by
mishing and mashing (including an extra mash at the end, because
that's how the rhythm goes!).

$ zig build run
before: 2934, after: 2424.

On to day 21!