Solve day 21 part 1.
A second constraint propagation problem for this year. I must be
getting used to Zig, because this was pretty straight-forward.
A bit boring maybe, but it's done.

I randomly picked my knowledge matrix to be indexed first by
ingredients, then by allergens, and I got the impression that
this was the more ergonomic choice. Yay!

$ zig build run
2262

On to part 2.