Solve day 21 part 2.
Internally the ingredients and allergens are given numbers, so I
didn't have to keep the original strings around after parsing, for
part 1. I'm glad I did, because now they're relevant!

This is the first answer this year which is really a string (as
opposed to, technically, a string representation of a number). And
like C, Zig makes you care about memory allocations, so that strings
are not really value types, and they're a bit more hassle to deal
with than numbers.

The puzzle text seems to suggest that part 1 could have been done
without actually figuring out which ingredient is which allergen,
but I don't see it.

$ zig build run
cxsvdm,glf,rsbxb,xbnmzr,txdmlzd,vlblq,mtnh,mptbpz

On to day 22.