Solve day 22 part 2.
A complete redesign of the game: since we're making copies of arrays
all over the place to keep a history of each game state, there's no
need to be cute with a circular buffer anymore.

Each sub-game uses its own ArenaAllocator to allocate copies of decks
of cards, then frees the whole arena once the sub-game is done. As a
small wrinkle, the final deck of cards from the sub-game returned to
the parent is allocated globally, and the parent is responsible for
freeing it.

Since `game()` calls itself recursively, I had to specify the set of
possible errors it returns explicitly, which seems reasonable.

I guess I was very careful when transcribing the rules into code,
because once the syntax errors were fixed, the program gave the right
answer on the first run (after a suspenseful 10 seconds of wondering
whether I accidentally put in an infinite loop).

$ zig build
$ time ./zig-cache/bin/aoc-2020
31657

real	0m10.601s
user	0m9.722s
sys	0m0.877s

On to day 23!