Solve day 23 part 1.
I love permutions, so I encoded the circular arrangement of cups into
a one-cycle permutation, where each cup points to the next cup in
clockwise order. Also I worked out that what the crab does amounts to
composing this cycle with a pair of swaps, which is conveniently a
constant-time operation.

Also, I allocated the permutation array with an extra element at the
front just so I could do 1-based indexing, but then it turned out
that I could use that extra slot as a temporary slot for swapping.

$ zig build run
75893264

On to part 2.