Solve day 24 part 2.
Haha another cellular automaton, augh. Oh well.

This time, for fun, I tried to do it in a map-reduce style, so for
each day:

1. Iterate over the black tiles and emit the coordinates of
   each neighbouring tile.

2. Sort this list of to collect repetitions next to each other.

3. Iterate over the result to get the count of black neighbours.
   (Actually, iterate over the original list in parallel to also
   know whether the count is for a black tile or a white tile.)

This relies on the fact that tiles with no black neighbours always
end up white the next day (because these tiles are never considered
in Step 3).

$ zig build run
3700

Given the small numbers of black tiles involved, using a sparse
representation like this was a fine choice.

On to day 25!